requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# LlamaIndex core and integrations (Stage 3+)
llama-index-core>=0.11.0
llama-index-vector-stores-qdrant>=0.3.0

# Embedding Backends
llama-index-embeddings-huggingface>=0.3.0  # HuggingFace/SentenceTransformers
llama-index-embeddings-ollama>=0.3.0       # Ollama
llama-index-embeddings-openai>=0.2.0       # For LM Studio (OpenAI-compatible)
sentence-transformers>=2.2.0               # Required for HuggingFace embeddings
torch>=2.0.0                               # Required for HuggingFace embeddings

# LLM Backends
llama-index-llms-ollama>=0.3.0            # Ollama
llama-index-llms-lmstudio>=0.2.0          # LM Studio

# LangGraph and LangChain (Stage 5+)
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# Data validation and modeling
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# HTTP client for Ollama API
httpx>=0.24.0

# JSON handling and utilities
ujson>=5.8.0

# Qdrant client
qdrant-client>=1.7.0

# Testing
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-asyncio>=0.21.0

# Code quality
black>=24.0.0
flake8>=7.0.0
mypy>=1.5.0

# Monitoring and logging
colorlog>=6.7.0

